Cleaned the datasets
The datasets are now organized in the column order discussed. I got rid of the "source" column as each dataframe was in its own section. The date is dtype.datetime. I created a list of the dataframes for later use if necessary.  I removed the sports scored from AP. Limited the df's to 1980 and newer. Searched for null values and found none. Looked for unique units in 'text' and 'headline'(how I got rid of the sports scores). Found some discrepancies in the characters but could not figure out how alter big blocks of text like this. luckily there weren't as many as I thought there would be. I'm sure there is more cleaning that can be done but that's all I got tonight.